+++ README.md
@@ -1,3 +1,5 @@
-[![Build Status](https://travis-ci.org/mkli90/wordspot.svg?branch=master)](https://travis-ci.org/mkli90/wordspot)
-[![Code Health](https://landscape.io/github/mkli90/wordspot/master/landscape.svg?style=flat)](https://landscape.io/github/mkli90/wordspot/master)
-[![Python Versions](https://img.shields.io/badge/python-3.5-blue.svg)](https://travis-ci.org/mkli90/wordspot)
+# Part-Structured Spotting
+
+[![Build Status](https://travis-ci.org/mkli90/pss.svg?branch=master)](https://travis-ci.org/mkli90/pss)
+[![Code Health](https://landscape.io/github/mkli90/pss/master/landscape.svg?style=flat)](https://landscape.io/github/mkli90/pss/master)
+[![Python Versions](https://img.shields.io/badge/python-3.5-blue.svg)](https://travis-ci.org/mkli90/pss)
