+++ setup.cfg
@@ -6,7 +6,7 @@ cover-inclusive=1
 cover-branches=1
 cover-erase=1
 cover-min-percentage=100
-cover-package=wordspot
+cover-package=pss
 
 [build_sphinx]
 source-dir = doc/source
